@@ -1,9 +1,10 @@
 package com.geterdone.android.todo.data;
 
-import android.arch.persistence.room.Database;
-import android.arch.persistence.room.Room;
-import android.arch.persistence.room.RoomDatabase;
 import android.content.Context;
+
+import androidx.room.Database;
+import androidx.room.Room;
+import androidx.room.RoomDatabase;
 
 @Database(entities = {Task.class}, version = 1)
 public abstract class TaskRoomDatabase extends RoomDatabase
